+++ CHANGELOG.rst
@@ -4,6 +4,13 @@ Changelog
 0.15.0 (unreleased)
 *******************
 
+Features:
+
+* Add new ``FlaskConfig()`` field type, to automatically serialise values from the
+  :ref:`Flask configuration object <flask:config>`.
+
+Other changes:
+
 * Only support Python>=3.6, marshmallow>=3.0.0, and marshmallow-sqlalchemy>=0.24.0.
 
 0.14.0 (2020-09-27)